Move update logic to own module
